refactor: move foodphotoflow out of its own dir
